pyproject.toml
[project]
name = "comicstrip-tutor"
version = "0.1.0"
description = "ComicStrip Tutor: storyboard-first technical teaching comics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-genai>=1.30.0",
  "jinja2>=3.1.6",
  "openai>=1.108.0",
  "pillow>=11.3.0",
  "pydantic>=2.11.9",
  "python-dotenv>=1.1.1",
  "rapidfuzz>=3.14.1",
  "rich>=14.1.0",
  "typer>=0.17.4",
]

[project.scripts]
comic-tutor = "comicstrip_tutor.cli:app"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.13.0",
  "ty>=0.0.1a18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/comicstrip_tutor"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
